almost done renovations
