+++ artemis/fileman/experiment_record.py
@@ -9,11 +9,11 @@ from artemis.general.test_mode import is_test_mode, TestMode
 import os
 from IPython.core.display import display, HTML
 from artemis.fileman.local_dir import format_filename, make_file_dir, get_local_path, get_relative_path
-from artemis.plotting.notebook_plots import show_embedded_figure
+from artemis.notebooks.notebook_plots import show_embedded_figure
 from artemis.fileman.notebook_utils import get_local_server_dir
 from artemis.fileman.notebook_utils import get_relative_link_from_relative_path
 from artemis.fileman.persistent_print import capture_print, stop_capturing_print
-from artemis.plotting.saving_plots import clear_saved_figure_locs, get_saved_figure_locs, \
+from artemis.notebooks.saving_plots_deprecated import clear_saved_figure_locs, get_saved_figure_locs, \
     set_show_callback, always_save_figures, show_saved_figure
 import matplotlib.pyplot as plt
 import re
