+++ artemis/plotting/pyplot_plus.py
@@ -25,55 +25,4 @@ def axvlines(xs, **specs):
     return [plt.axvline(x, **specs) for x in xs]
 
 
-_SHOW_FCN = plt.show
 
-
-def draw():
-    plt.draw()
-    plt.pause(0.00001)
-
-
-_SHOW_AND_CONTINUE_FCN = draw
-
-
-def set_show_and_hang_callback(cb):
-    """
-    :param cb: A function that takes an optional figure as an argument.  Or None to reset to the original show.
-    """
-    global _SHOW_FCN
-    if cb is None:
-        _SHOW_FCN = plt.show
-    else:
-        _SHOW_FCN = cb
-
-
-def set_show_and_continue_callback(cb):
-    global _SHOW_AND_CONTINUE_FCN
-    if cb is None:
-        _SHOW_AND_CONTINUE_FCN = draw
-    else:
-        _SHOW_AND_CONTINUE_FCN = show_and_continue
-
-
-def show_and_hang():
-    _SHOW_FCN()
-
-
-def show_and_continue():
-    _SHOW_AND_CONTINUE_FCN()
-    # pause(0.00001)  # Required to display, at least on some backends.
-
-
-class SaveToFile():
-
-    def __init__(self, path):
-        self._path = get_local_path(path)
-        self._first = False
-
-    def __call__(self, fig = None):
-        if self._first:
-            ARTEMIS_LOGGER.info('Creating figure at "%s"' % (self._path, ))
-            self._first = False
-        if fig is None:
-            fig = plt.gcf()
-        fig.savefig(self._path)
